api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.52
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
psycopg2-binary==2.9.5